+++ src/builtins/alias.rs
@@ -8,7 +8,7 @@ use crate::types::Tokens;
 
 pub fn run(sh: &mut shell::Shell, tokens: &Tokens) -> i32 {
     if tokens.len() == 1 {
-        return show_alias_list(sh, "");
+        return show_alias_list(sh);
     }
     if tokens.len() > 2 {
         println_stderr!("alias syntax error");
@@ -27,7 +27,7 @@ pub fn run(sh: &mut shell::Shell, tokens: &Tokens) -> i32 {
         }
     }
     if re_single_read.is_match(input) {
-        return show_alias_list(sh, input);
+        return show_single_alias(sh, input);
     }
 
     let re_to_add;
@@ -47,18 +47,22 @@ pub fn run(sh: &mut shell::Shell, tokens: &Tokens) -> i32 {
     0
 }
 
-fn show_alias_list(sh: &shell::Shell, name_to_find: &str) -> i32 {
-    let mut single_and_not_found = true;
+fn show_alias_list(sh: &shell::Shell) -> i32 {
     for (name, value) in sh.get_alias_list() {
-        if name_to_find.is_empty() {
-            single_and_not_found = false;
+        println!("alias {}='{}'", name, value);
+    }
+    0
+}
+
+fn show_single_alias(sh: &shell::Shell, name_to_find: &str) -> i32 {
+    let mut found = false;
+    for (name, value) in sh.get_alias_list() {
+        if name_to_find == name {
             println!("alias {}='{}'", name, value);
-        } else if name_to_find == name {
-            println!("alias {}='{}'", name, value);
-            single_and_not_found = false;
+            found = true;
         }
     }
-    if single_and_not_found {
+    if !found {
         println_stderr!("cicada: alias: {}: not found", name_to_find);
         return 1;
     }