+++ src/builtins/mod.rs
@@ -9,4 +9,5 @@ pub mod fg;
 pub mod history;
 pub mod jobs;
 pub mod source;
+pub mod unalias;
 pub mod vox;