+++ src/completers/path.rs
@@ -172,7 +172,7 @@ fn complete_bin(sh: &shell::Shell, path: &str) -> Vec<Completion> {
     }
     let builtins = vec![
         "alias", "bg", "cd", "cinfo", "exec", "exit", "export", "fg", "history",
-        "jobs", "source", "vox",
+        "jobs", "source", "unalias", "vox",
     ];
     for item in &builtins {
         if !item.starts_with(fname) {